// frontend/src/sections/supplier-returns/preferences.generated.ts
/* eslint-disable */
// THIS FILE IS GENERATED by codegen/plugin.js — do not edit by hand.
import type { TypedDocument } from "../../api/graphql";

export type SupplierReturnPreferencesVariables = {
  storeId: string;
};

export type SupplierReturnPreferencesResult = {
  preferences: {
  disableManualReturns: boolean;
  invoiceStatusOptions: Array<"NEW" | "ALLOCATED" | "PICKED" | "SHIPPED" | "DELIVERED" | "RECEIVED" | "VERIFIED" | "CANCELLED">;
};
};

export const SupplierReturnPreferences = {
  query: "query supplierReturnPreferences($storeId: String!) {\n  preferences(storeId: $storeId) {\n    disableManualReturns\n    invoiceStatusOptions\n  }\n}",
} as TypedDocument<SupplierReturnPreferencesResult, SupplierReturnPreferencesVariables>;